BAH-4858 | Refactor. Extract out extractFormFieldPath and extractFormFieldPath in observations utils
