+++ packages/bahmni-widgets/src/forms/FormsTable.tsx
@@ -24,11 +24,11 @@ import { useQuery, useQueryClient } from '@tanstack/react-query';
 import { Bundle, Observation } from 'fhir/r4';
 import React, { useCallback, useMemo, useState } from 'react';
 import { usePatientUUID } from '../hooks/usePatientUUID';
+import { extractFormFieldPath } from '../observations/utils';
 import { ObservationsRenderer } from '../observationsRenderer';
 import { WidgetProps } from '../registry/model';
 import { FormRecordViewModel, GroupedFormRecords } from './models';
 import styles from './styles/FormsTable.module.scss';
-import { extractFormFieldPath } from './utils';
 
 interface FormsTableConfig {
   numberOfVisits?: number;
