+++ packages/bahmni-widgets/src/tasks/views/ViewFormData/ViewFormData.tsx
@@ -9,12 +9,12 @@ import {
 import { useQueries, useQuery } from '@tanstack/react-query';
 import type { Bundle, Encounter, Observation } from 'fhir/r4';
 import React, { useMemo } from 'react';
+import { extractFormFieldPath } from '../../../observations/utils';
 import { ObservationsRenderer } from '../../../observationsRenderer';
 import type { TaskView, TaskViewModel } from '../../models';
 import { extractFormNameFromTask } from '../../utils';
 import styles from './ViewFormData.module.scss';
 import {
-  extractFormFieldPath,
   extractUuidFromReference,
   groupObservationsByEncounter,
   type EncounterGroup,
